Create methods to determine word count.
